Part 1 of refactoring frame-commandbuffer-etc. stuff
 * You can now get from a `DeviceContext` to its `Device`. Many places
   where we used to have to use `g_hackActiveThreadContext` should no
   longer need it, but most of them haven't been changed. (And there
   are still many other places where we do still need it anyway.)

 * New `IDevice::Stall` method that blocks until the GPU has finished
   with (at least) everything before the current frame. All of the
   more complicated methods that did this plus other stuff, like
   `ThreadContext::WaitUntilQueueCompleted` and `SyncEventSet::Stall`,
   are gone. We never need all of that complicated stuff. We _do_
   sometimes need to `EndHeadlessFrame` before or after a stall, but
   we should (and now do) call that explicitly.

 * Every unit test but one that should pass now does, on Apple Metal
   and GL, OSX and iOS.

 * Various little fixes that were prototyped on the
   `-graphicspipeline` branch redone on this branch (and mostly
   simpler with the `Device` changes.

Part 2 will be the changes that ensure that a `DeviceContext` always
has a current frame, and a current command buffer, on targets where
that matters; there's a public/cross-API way to say "commit a bunch of
non-drawing commands" so Apple Metal's `[Begin|End]HeadlessFrame` can
die; and `ConstantBuffer` works properly and consistently across
targets in both synchronized and unsynchronized mode. At which point
the last broken unit test will pass, and can be expanded to test more
stuff.