Make IEntityInterface::PrintDocument a pure virtual function.
That function was a regular virtual function implemented in EntityInterface.cpp. The problem is that the cocos3d Xcode project don't compile EntityInterface.cpp at all, which causes a linking issue. Making it a pure virtual function will get rid of the linking problem but will most likely break compilation of EntityInterface.cpp (EntityInterface.vcxproj seems to be the only project that uses it).

Found this while compiling ParticleEditor.